Added dash!
Added the dash, removed the old "hold shift to move fast", you can now press the dash key to rapidly move in a direction at the cost of having to wait just a bit right after dashing. Once the health and damage system works, ill add some i-frames to the start of this dash (already kind of coded it in, if the startDash value is greater then 0 then you would be immortal), will also add a limit to dash based on stamina bar.